Initial commit
This is a working prototype, without any UI.

It just searches for the name of the resource under the respective resource type on the other service. The obviously gives the wrong results sometimes. For albums and songs there is a lot of room for improvement through validating other properties of the resouce.

I won't publish to the chrome store until those low hanging fruits are picked.

There is the [declerativeWebRequest](https://developer.chrome.com/extensions/declarativeWebRequest) API but it's only in beta and I'm not sure if it's comming soon but it would be worth looking into once it's stable.